finition de la sauvegarde (a tester quand le joueur aura des cartes)

+++ source/espacejeux.hpp
@@ -33,7 +33,7 @@ class Pyramide {
         const Carte& acheterCarte(int i, int j); // retourne la carte de la pyramide et la supprime + rempli la case avec une nouvelle carte de la pioche
         const Carte& ReserverCartePioche(int niveau); //retourne une carte de la pioche
         
-        int getNbCartesNiv(int niveau) const {return array_cartes[niveau].size(); }; // retourne le nombre de cartes d'un niveau dans la pyramide
+        size_t getNbCartesNiv(int niveau) const {return array_cartes[niveau].size(); }; // retourne le nombre de cartes d'un niveau dans la pyramide
         void afficherPyramide() const; // affiche la pyramide
         
         Pioche& getPioche1() const { return pNv1; }
